refactor: separando os dto's por contexto
