+++ src/main/java/br/com/solutis/locadora/model/dto/car/AccessoryDto.java
@@ -1,5 +1,6 @@
-package br.com.solutis.locadora.model.dto;
+package br.com.solutis.locadora.model.dto.car;
 
+import br.com.solutis.locadora.model.dto.AbstractDto;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
@@ -8,7 +9,7 @@ import lombok.EqualsAndHashCode;
 
 @Data
 @EqualsAndHashCode(callSuper = true)
-public class CarAccessoryDto extends AbstractDto {
+public class AccessoryDto extends AbstractDto {
     @NotNull(message = "Description is required")
     @NotBlank(message = "Description is required")
     @Size(min = 1, max = 255, message = "Description must be between 1 and 255 characters long")
