+++ src/main/java/br/com/solutis/locadora/model/dto/car/CarDto.java
@@ -1,5 +1,6 @@
-package br.com.solutis.locadora.model.dto;
+package br.com.solutis.locadora.model.dto.car;
 
+import br.com.solutis.locadora.model.dto.AbstractDto;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
