+++ src/main/java/br/com/solutis/locadora/model/dto/car/ModelDto.java
@@ -1,5 +1,6 @@
-package br.com.solutis.locadora.model.dto;
+package br.com.solutis.locadora.model.dto.car;
 
+import br.com.solutis.locadora.model.dto.AbstractDto;
 import br.com.solutis.locadora.model.entity.car.ModelCategoryEnum;
 import jakarta.persistence.EnumType;
 import jakarta.persistence.Enumerated;
@@ -11,7 +12,7 @@ import lombok.EqualsAndHashCode;
 
 @Data
 @EqualsAndHashCode(callSuper = true)
-public class CarModelDto extends AbstractDto {
+public class ModelDto extends AbstractDto {
     @NotNull(message = "Description is required")
     @NotBlank(message = "Description is required")
     @Size(min = 1, max = 255, message = "Description must be between 1 and 255 characters long")
