+++ src/main/java/br/com/solutis/locadora/model/dto/person/EmployeeDto.java
@@ -1,4 +1,4 @@
-package br.com.solutis.locadora.model.dto;
+package br.com.solutis.locadora.model.dto.person;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
