+++ src/main/java/br/com/solutis/locadora/service/person/DriverService.java
@@ -1,7 +1,7 @@
 package br.com.solutis.locadora.service.person;
 
 import br.com.solutis.locadora.mapper.person.DriverMapper;
-import br.com.solutis.locadora.model.dto.DriverDto;
+import br.com.solutis.locadora.model.dto.person.DriverDto;
 import br.com.solutis.locadora.model.entity.person.Driver;
 import br.com.solutis.locadora.repository.CrudRepository;
 import br.com.solutis.locadora.service.CrudService;
