+++ src/main/java/br/com/solutis/locadora/service/rent/InsurancePolicyService.java
@@ -2,7 +2,7 @@ package br.com.solutis.locadora.service.rent;
 
 import br.com.solutis.locadora.exception.BadRequestException;
 import br.com.solutis.locadora.mapper.rent.InsurancePolicyMapper;
-import br.com.solutis.locadora.model.dto.InsurancePolicyDto;
+import br.com.solutis.locadora.model.dto.rent.InsurancePolicyDto;
 import br.com.solutis.locadora.model.entity.rent.InsurancePolicy;
 import br.com.solutis.locadora.repository.CrudRepository;
 import br.com.solutis.locadora.service.CrudService;
